web api had a potential race if I was deleting moose in very rare cases
this change breaks the moose db contract, but by design to make it less problematic.
the main purpose of the trait is to allow for different backing dbs in the future.

TODO: All tasks, like the db dump and import tasks, need to be represented by the MooseDB trait.
Other: All release builds will abort() on panic from now on.
       This is to make sure the application dies if unexpected database (or other logic errors) happen.
       Dying is better than pretending the application can gracefully continue in the face of these issues.